% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{f2}
\alias{f2}
\title{Function calculates either the mean, median, or standard deviation of the DRG
codes. over all of the DRG codes for average Medicare payments}
\usage{
f2(statistics)
}
\arguments{
\item{statistics}{is a string which specifies what statistic is needed}
}
\value{
returns the chosen statistic calculated
}
\description{
Function calculates either the mean, median, or standard deviation of the DRG
codes. over all of the DRG codes for average Medicare payments
}
\examples{
f2("mean")

}
